+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-06 21:13:39
-// Commit: fix(test): prevent double tap duplicate action in test
+// Updated at: 2026-08-08 14:25:07
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '3427';
-  final String lastMessage = 'fix(test): prevent double tap duplicate action in test';
-  final DateTime timestamp = DateTime.parse('2026-08-06T21:13:39+05:00');
+  final String commitRef = '3468';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-08-08T14:25:07+05:00');
 }